Updated node-require to take non-whole numbers

+++ node-require/calculate.js
@@ -3,8 +3,9 @@ const subtract = require('./subtract');
 const multiply = require('./multiply');
 const divide = require('./divide');
 
-var firstNum = parseInt(process.argv[2]);
-var secondNum = parseInt(process.argv[4]);
+var firstNum = parseFloat(process.argv[2]);
+var secondNum = parseFloat(process.argv[4]);
+
 if (process.argv[3] === 'plus') {
   console.log('result:', add(firstNum, secondNum));
 } else if (process.argv[3] === 'minus') {
